Modified service to save notebook submission and post grades to EdX.
This commit also encapsulates some design changes to the user
experience. Because of future grade disputes, keeping a copy of the
user's submission is important. Now the gofer service will save a
timestamped version of the submitted notebooks before grading. The
notebook extension was modified to ease any user concerns about their
submission not being saved, alerting immediately after submission but
not displaying the notebook's grade in the browser.

Grades are posted back to EdX with the help of a config file that
contains the lab or assignment ids and outcomes_url (a gradebook link).